feat(intex-contracts): rate bids, entry anchor, call params on the auction wire
Mirror the Rust desis parameter-model rework on the Solidity (BNB) side so both
chains agree on the AUCTION_STAGE_START and BIDS_BATCH wire. The bid is now a
rate (% of strike) end to end: revealed as a rate, escrowed against an
entry-derived strike, relayed and cleared as a rate.

- BridgeMsgCodec: AUCTION_STAGE_START body grows to 72 bytes carrying
  minIntexBidRate, entryPrice, floorPriceMinor, callPriceMinor and the call
  trigger (window/threshold/period), dropping costAmountMinor; BIDS_BATCH
  carries intexBidRates (uint32) instead of prices; AUCTION_RESULT clearing
  field renamed clearingRate; add RATE_SCALE = 1e6.
- Origin/Target messengers + IDesis/IOriginMessenger/ITargetMessenger: plumb
  the new fields; TargetMessenger forwards entry/rate/call into auctionStart.
- IntexAuction: AuctionParams holds entry/rate/call; bids are a 1e6 fixed-point
  rate; escrow lock = qty * strike * rate / RATE_SCALE with strike =
  entryPrice * PROMIS_LOAD / 1e12 rounded up to 100 (bit-for-bit with desis
  from_entry_price, so finalize never reverts on a width mismatch); EIP-712
  RevealBid signs bidRate; clearing floor checks clearingRate.

ISSUANCE_INSTRUCTIONS wire is left unchanged (the sender still ships
costAmountMinor), so SeriesData and createSeries are untouched.

forge test: 587 passed.

+++ contracts/intex/test/mocks/MockDesis.sol
@@ -25,7 +25,7 @@ contract MockDesis {
         uint32, /* relayGeneration */
         address[] calldata, /* bidderAddresses */
         uint16[] calldata, /* intexQuantities */
-        uint64[] calldata, /* intexBidPrices */
+        uint32[] calldata, /* intexBidRates */
         uint32[] calldata /* timestamps */
     ) external {}
